Delay ODL startup and disable console
Delay ODL startup until networking service is up, otherwise if all
interfaces on the box come up with DHCP and there is no local
definition of FQDN (as in `/etc/hosts`), ODL fails to startup.

Secondly disable the server console when starting up, otherwise it will
hijack the startup tty.

Change-Id: I2e973d4207f95e1ef823c269aaa82b3b1fe5b71d
Signed-off-by: Daniel Farrell <dfarrell@redhat.com>