Update templates and tarball URL for Li SR3
SR3 replaced SR2 as the latest Li RPM on the CBS, so the 3-candidate
repo config started pulling 3.3 and using 3.2 templates. This is a quick
fix, but the desired longer-term solution is to add a param to the ODL
Puppet module to allow for release selection.

CentOS 7 Docker Beaker tests are passing.

Change-Id: Ib001850a783ee6de071d8a8bb1962dca4898a958
Signed-off-by: Daniel Farrell <dfarrell@redhat.com>