Add param to config RPM repo to install ODL from
Param allows user to change which version of ODL installed without
changes to puppet-opendaylight. Even better than limiting to a specific
version, the new rpm_repo param allows the user to select a the latest
RPM from a major release or SR, or a fixed and official release.

This patch includes the feature changes, rspec-puppet test changes,
Beaker test changes, in-code docs and README docs.

Change-Id: Ie01dafd485318d98410a2847e53ebd819a5cf7c3
Signed-off-by: Daniel Farrell <dfarrell@redhat.com>