Adds ability to pass Java options to ODL startup
Previously we always set the Java options to a single value which will
only run Java with ipv4 stack.  There is a requirement to be able to set
other parameters like heap seize or enable garbage collection.  This
change exposes a parameter where many Java options can be set via a
string.

Change-Id: Ib79a83dd59f65a61912a4917e18e41196e7f91cd
Signed-off-by: Tim Rozet <tdrozet@gmail.com>
Signed-off-by: Daniel Farrell <dfarrell@redhat.com>