Allow for disabling repository configuration
This change moves the repository configuration to it's own class so it
can be included externally and excludeded if necessary. As part of this
change, the inclusion of rspec-puppet-facts was done to simplify testing
support os configurations.
See https://github.com/mcanevet/rspec-puppet-facts for details on how to
work with rspec-puppet-facts. By default it uses the metadata.json for
the supported operating systems. Since 14.04 is not a supported
configuration based on the failure in ::opendaylight, this has also been
updated to reference 16.04 as a supported version.

Change-Id: Iae079ba98c6f2a63e5c4d0cc11c2ffb8cdca0515
Signed-off-by: Daniel Farrell <dfarrell@redhat.com>