Force puppetlabs-apt version 2.4, 3.0 breaks rspec
The upgrade to 3.0 started causing our daily CI to fail and is now
causing per-patch test-sanity job to fail. I think it has to do with
the apt module moving to "data in modules", vs params.

Not sure how to actually fix the problem. Also not sure if the apt
devs meant to break everyone like this. Will raise a ticket.

Change-Id: I6f6996ab392abbe8ddbe40e3b8221a650bc9ed8a
Signed-off-by: Daniel Farrell <dfarrell@redhat.com>