Split tests into groups based on ODL major version
Per Tim, we need to move to branched model for puppet-opendaylight. Need
to support Boron in a stable/boron branch but allow it to break on
master or stable/carbon. New username/password change for example works
in Carbon but will never work in Boron becuase the required JAR will not
be backported.

Once we have split branches, we can refactor the test code to use the
right default ODL version per-branch. For now, just modify the tests we
call from master to only include Carbon tests.

Change-Id: I894e8aed644e2d9c5d8c966ac4a72c5a28578606
Signed-off-by: Daniel Farrell <dfarrell@redhat.com>