Disables modifying password of user
Until we have support for detecting if an existing user's password
matches the desired password, we need to disable support for setting it.
The current behavior is we don't know what the current password is so we
always configure it.  This means our puppet module is not idempotent and
triggers ODL to restart everytime puppet is run, which is bad.

The request to add support for the idm tool to be able to check password
is:  https://bugs.opendaylight.org/show_bug.cgi?id=8721

Note this does not affect odl usernames.  We still detect if a user
already exists and if not create it with the desired password.

Change-Id: I64a23ad6fe68d1afb765f46c6b54f7291d25b47e
Signed-off-by: Tim Rozet <trozet@redhat.com>