Fixes idempotency and maintence with logging
ODL will modify the loggin file at run time to add qos and other
configs.  This causes a problem where when puppet runs again it uses a
template so it overrides the file and restarts ODL.  This change
abandons using a template and instead just makes sure each logger is set
in the file using file_line.

Change-Id: Idd7d565be61a73985203b64c200281c9242c87bb
Signed-off-by: Tim Rozet <trozet@redhat.com>