Fixes clustering to use template instead of exec
Adds ability to specify modules to use when breaking up the datastore
into shards.  By default we just use the 'default' module which contains
all of the datastore, but a user is able to pass a hash of modules and
namespaces if they desire to break up the datastore into separate
shards.

Deprecates any need to use ha_node_index param.

Change-Id: I05c88741c16c433baa11fe4cbc43bc04fc8c2625
Signed-off-by: Tim Rozet <trozet@redhat.com>