Bug 8854: Add config of ODL log file size/rollover
The default ODL logging config uses 10 small log files. This makes
searching for errors difficult. Add the ability to configure the maximum
size of karaf.log and the number of rollover logs to keep. Default to
values that make it normal to end up with a single large log file but
still allow pruning from the tail when logs rollover.

Change-Id: Ibcdd0c6634fc37180a9880e03abfd24f1eb84a10
Signed-Off-By: Janki Chhatbar <jchhatba@redhat.com>
Signed-off-by: Daniel Farrell <dfarrell@redhat.com>