Fix bind IP config for REST
Changes bind ip to be set as a default attribute rather than value and
also configures pax web cfg as recommended in
https://bugs.opendaylight.org/show_bug.cgi?id=9158

Note explicitly configuring 0.0.0.0 does not work so if default we skip
configuring bind IP and it defaults to 0.0.0.0.

Also increases sleep before curling REST in acceptance tests because
Karaf 4 takes longer to boot.

Change-Id: I23e7b1d501c782d5d1df0aa6d8b3e6cac9540de7
Signed-off-by: Tim Rozet <trozet@redhat.com>