Fixes to TLS ODL keystore provider
File exist checks were done in in the odl_keystore type.  Checks within
a type are done before puppet is applied.  Therefore the file has to
exist before puppet is ran, which excludes another puppet resource from
being able to create the files at run time.

The JKS provider for the keystore was incorrectly passing the CA
certificate as a file to create the p12 keystore, instead of as a X509
certificate object.

Changes Include:
 - Checks to ensure key files, ca cert, and public cert are moved from
   the type into the jks provider.
 - CA cert is now correctly created as and OpenSSL:X509 object before
   being passed into the p12 keystore create.

Change-Id: Iaf3e66fa44dee3987817177e2210230ee2d3325c
Signed-off-by: Tim Rozet <trozet@redhat.com>