Parametrize the logging pattern to use
Currently the packaging of opendaylight is being done with the current
opendaylight default. This default does not always suits the needs of
every deployment and it is good to have this configurable.

This patch adds a `log_pattern` configuration that allows to specify the
logging pattern that will be used. At the same time, it also defines a
default that is non dependant on PaxOsgi on being enabled [1]. The new
default removes the '%X{bundle.id} - %X{bundle.name} -
%X{bundle.version}' bits that are PaxOsgin related, and to compensate
for that information removed it adds more context to the class name and
expands the clipping of the column to 60 characters.

[1] https://bugzilla.redhat.com/show_bug.cgi?id=1650576

Change-Id: I923cd372c71ba5c04a27f6d4478ca7c511c14d2d
Signed-off-by: Michel Peterson <michel@redhat.com>
(cherry picked from commit a469ad6b5bd72bdd4fca39f1b87149d622d1975f)