Add boolean flag to disable the PaxOsgi appender
Currently the PaxOsgi appender is on the OpenDaylight default
configurations for the logging system. It was found, during a research
of performance issues [1], that this adds pressure to the Java GC and is
a facility that is interesting to have configurable.

This patch adds the configuraton boolean `enable_paxosgi_appender` that
allows packaging OpenDaylight with this setting disabled. In addition,
due to a business requirement [2], this setting is set to `false` by
default.

[1] ODLPARENT-176
[2] https://bugzilla.redhat.com/show_bug.cgi?id=1650576

Change-Id: I358618594a5703cc8a22575dde88e9273e2b8a16
Signed-off-by: Michel Peterson <michel@redhat.com>
(cherry picked from commit 965ff409ba70a1c8754e0f70b4b4b5b7b86bd428)